Refactor chat window component
